+++ policyengine_canada/variables/household/expenses/childcare/childcare_received_hours_per_day.py
@@ -1,0 +1,9 @@
+from policyengine_canada.model_api import *
+
+
+class childcare_received_hours_per_day(Variable):
+    value_type = int
+    entity = Household
+    label = "Child care hours per day"
+    unit = "hour"
+    definition_period = YEAR