+++ policyengine_canada/variables/household/expenses/childcare/full_time_childcare.py
@@ -1,0 +1,11 @@
+from policyengine_canada.model_api import *
+
+
+class full_time_childcare(Variable):
+    value_type = bool
+    entity = Household
+    label = "Full-time child care"
+    definition_period = YEAR
+
+
+# TODO: computation based variable - based on full-time definition